refactor(tests): update modexp upper bounds framework
